TcOCR/requirements.txt
ultralytics
transformers
torch
torchvision
opencv-python
Pillow
numpy
roboflow